+++ regimes/regimes.go
@@ -14,6 +14,7 @@ import (
 	_ "github.com/invopop/gobl/regimes/es"
 	_ "github.com/invopop/gobl/regimes/fr"
 	_ "github.com/invopop/gobl/regimes/gb"
+	_ "github.com/invopop/gobl/regimes/gr"
 	_ "github.com/invopop/gobl/regimes/it"
 	_ "github.com/invopop/gobl/regimes/mx"
 	_ "github.com/invopop/gobl/regimes/nl"