+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - `num.Amount` - added `RescaleDown` method, that helps reduce accuracy from a number if higher.
 - `num.Amount` - added `RescaleRange` method, ensures that the exponent is within a specific range.
+- Greece tax regime
 
 ## [v0.79.3] - 2024-06-18